fix: decoder API, rearrange files
